requirements.txt
llama-index==0.10.50
llama-index-agent-openai==0.2.7
llama-index-llms-openai==0.1.23
llama-index-embeddings-openai==0.1.10
llama-index-readers-file==0.1.25
llama-index-readers-json==0.1.5
openai==1.35.5
pdfplumber==0.11.1
streamlit==1.36.0
llama-index-program-openai==0.1.6
llama-index-response-synthesizers-google==0.1.3
python-dotenv==1.0.1
llama-parse==0.4.6
pytest==8.3.2
pytest-asyncio==0.23.8
pyvis==0.3.2
python-slugify==8.0.4
